requirements.txt
flask==3.0.0
numpy==1.24.3
pandas==2.1.0
pillow==10.0.0
matplotlib==3.7.2
scikit-learn==1.3.0
deep-translator==1.11.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
langdetect==1.0.9
opencv-python==4.8.0.76
werkzeug==3.0.0
torch==2.1.0
torchvision==0.16.0